+++ R/transform_methods.R
@@ -106,9 +106,11 @@ transform_pload.list_inning_all <- function(payload_obj, ...) {
                       event2 = if (exists('event2', where = payload_obj$pitch)) event2 else NA,
                       event2_es = if (exists('event2_es', where = payload_obj$pitch)) event2_es else NA,
                       # tfs and tfs_zulu columns may be blank for older data sets. If blank, set them to NA.
-                      tfs = ifelse(tfs == "", NA, tfs), tfs_zulu = ifelse(tfs_zulu == "", NA, tfs_zulu)) %>%
+                      tfs = ifelse(tfs == "", NA, tfs), tfs_zulu = ifelse(tfs_zulu == "", NA, tfs_zulu),
+                      # Some spring training and minor league games may be missing speed data.
+                      start_speed = ifelse(start_speed == "", NA, as.numeric(start_speed))) %>%
         
-        dplyr::mutate(id=as.numeric(id), x=as.numeric(x), y=as.numeric(y), start_speed=as.numeric(start_speed), 
+        dplyr::mutate(id=as.numeric(id), x=as.numeric(x), y=as.numeric(y), 
                       end_speed=as.numeric(end_speed), sz_top=as.numeric(sz_top), sz_bot=as.numeric(sz_bot),
                       pfx_x=as.numeric(pfx_x), pfx_z=as.numeric(pfx_z), px=as.numeric(px), pz=as.numeric(pz),
                       x0=as.numeric(x0), y0=as.numeric(y0), z0=as.numeric(z0), vx0=as.numeric(vx0), 
@@ -155,7 +157,25 @@ transform_pload.list_linescore <- function(payload_obj, ...) {
                                         inning=as.numeric(inning), balls=as.numeric(balls), strikes=as.numeric(strikes),
                                         outs=as.numeric(outs), away_team_runs=as.numeric(away_team_runs), home_team_runs=as.numeric(home_team_runs),
                                         away_team_hits=as.numeric(away_team_hits), home_team_hits=as.numeric(home_team_hits), 
-                                        away_team_errors=as.numeric(away_team_errors), home_team_errors=as.numeric(home_team_errors))
+                                        away_team_errors=as.numeric(away_team_errors), home_team_errors=as.numeric(home_team_errors)) %>%
+        
+        dplyr::select(ampm, aw_lg_ampm, away_ampm, away_code, away_division, away_file_code,       
+                      away_league_id, away_loss, away_name_abbrev, away_preview_link, away_recap_link, away_sport_code,      
+                      away_team_city, away_team_errors, away_team_hits, away_team_id, away_team_link, away_team_name,       
+                      away_team_runs, away_time, away_time_zone, away_win, balls, day,                  
+                      description, double_header_sw, first_pitch_et, game_data_directory, game_nbr, game_pk,              
+                      game_type, gameday_link, gameday_sw, highlights_available, hm_lg_ampm, home_ampm,           
+                      home_code, home_division, home_file_code, home_league_id, home_loss, home_name_abbrev,     
+                      home_preview_link, home_recap_link, home_sport_code, home_team_city, home_team_errors, home_team_hits,       
+                      home_team_id, home_team_link, home_team_name, home_team_runs, home_time, home_time_zone,       
+                      home_win, id, ind, inning, inning_break_length, inning_state,         
+                      is_no_hitter, is_perfect_game, league, location, note, original_date,        
+                      outs, pbp_last, preview, reason, runner_on_base_status, scheduled_innings,    
+                      status, strikes, tbd_flag, tiebreaker_sw, time, time_aw_lg,           
+                      time_date, time_date_aw_lg, time_date_hm_lg, time_hm_lg, time_zone, time_zone_aw_lg,      
+                      time_zone_hm_lg, top_inning, tz_aw_lg_gen, tz_hm_lg_gen, venue, venue_id,             
+                      venue_w_chan_loc, wrapup_link, xmlns.xs, date, runner_on_2b, runner_on_3b, runner_on_1b)
+    
     return(payload_obj)
 }